JVMCBC-870: fix completion of view results with reduce enabled
A bug in the chunk response parser prohibited responses that
came from reduced view results to be completed and as a result
timed out on the user side.

A regression test has been added to verify this functionality.

Note: because of an edge case in views with like limit(0)
and reduce(true) only an empty rows list is emitted, I had
to add one more clue to the header check to assert if it is
the last chunk or not, last chance to emit the header.
This is right now only used with views, the other parsers
do not have a similar problem so just ignore the property.

Change-Id: I8a8b3f77e8dc1829d98d816c0b0fcff4286d9759
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/133619
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/133628
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>