JVMCBC-874: Distinguish outdated manifest vs. collection not found
It is a little confusing that the same retry reason is used if the
server tell us the manifest is outdated vs. the collection we
are looking for is not even in the manifest we own.

In the latter case, it makes sense to have a new specific retry
reason (collection not found) so that it is clear what is going on.
It does not change the behavior of the op (it might still time out
eventually), but now it should be more clear from the retry reason
why it timed out.

Change-Id: I97c131991c11fc9dc172d01b3e8895ad0327f7cd
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/134400
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>