JVMCBC-875: Expose RetryReasons on TimeoutException
It is very useful to have the retry reasons and number of retry
attempts available directly on the exception when trying to apply
additional checks/logic in the user level code.

They have been accessible through a couple indirections already,
this just makes it more convenient from an API usage point of view.

Side note: this also removes the timeout for a recently added test
so it runs a bit more quickly.

Change-Id: I668d45faa81e4b8d0fcf2cc8fa2b7c1ffd22ce47
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/134401
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/135149
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>