JCBC-1696: Allow withExpiry and custom transcoder
This changeset makes sure that even when a custom transcoder
is asked for on withExpiry, it doesn't break. Previously we would
assume JSON flags, which might not match up with a custom transcoder
being used to store the value.

If the expiry is asked for, we also fetch the flags via subdoc and
pass it down to the get result so that the custom transcoder is not
thrown off by the wrong flags.

Note that along the way I realized this code was duplicating the
flags from the lookup in enum, so that has been cleaned up in a
drive-by fashion.

Change-Id: I2125cf03e8c39f74463405b64cefc1a20720a96b
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/134672
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/135151
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>