JVMCBC-880: Track concurrent open bucket attempts for bucketConfigLoadInProgress.
While this title might sound a little complicated, it's pretty simple what is going
on: when more than  one openBucket attempts (i.e. if you open 5 buckets) in parallel,
the first one will reset bucketConfigLoadInProgress to false even though the other
4 are still going on.

Switching to a reference count (of sorts) we can better track the concurrent attempts
and only return false if the counter is at 0.

I had to modify the provider so that parts can be mocked out for the tests, although
I think the refactoring itself also makes the code clearer so a win-win hopefully.

Change-Id: Ide249bd1182cbfa393b564d707ea27a675913781
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/134998
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>