JVMCBC-882: Prevent disabing KV service prematurely on node
Motivation
----------
During the investigation of bootstrapping against non-kv nodes
it has been found that sometimes that initial bootstrap service/endpoint
keeps reconnecting and is not cleaned up properly after a config
could be located and is applied to the core.

Modifications
-------------
The issue happens because we were prematurely setting the 1 bit
back to 0 for the kv service in the "servicesEnabled" cache. This
can happen if i.e. the gcccp kv service and a bucket-level kv service
were both active on the same node and the first one got cleaned up,
setting it to 0 and then when the second one is supposed to be cleaned
up it got filtered at the core level since serviceEnabled was already
reporting false.

This in turn prevented it from being disconnected and lingering around
for longer than it actually needed to.

Result
------
Proper, non-racy, cleanup of kv sockets when bootstrapping against
a non-kv node post 6.5.

Change-Id: Ib422ae3b38dd2c938cc9f994143ccf6a1f7e3932
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/135172
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>