JVMCBC-883: Be less verbose about a DNS SRV timeout
Right now if DNS SRV times out we raise a warning with an exception.
In some cases this might be okay, but in most cases the user just
wants to bootstrap against the actual hostname and in this case
this is a wrong warning and misleading/noisy.

This changeset dials it back to a info level message which can
still be analyzed and with no noisy stack trace. If the intent
of the user was to bootstrap with DNS SRV and it timed out, the
subsequent bootstrap attempt will also fail and generate warnings
anyways.

Change-Id: Ib38442484cbc1dc08409b17c5d1ea3f7fc01258d
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/135175
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
(cherry picked from commit 38a3b188e042a720f11e774a09001ea6ca06a0a4)
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/135109
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>