JVMCBC-886: Improve SASL auth failures with LDAP enabled
Motivation
----------
The server changed the details on which SASL mechanism it
sends when LDAP is enabled a few times in the past. 6.5 will
only send PLAIN, earlier and later releases will send all
of them. The revert has been made by intention though since
other components were found non compatible with the 6.5
approach.

Since LDAP only supports PLAIN, and the SDK has no way of
knowing upfront if LDAP is enabled, the user must manually
change it on the PasswordAuthenticator.

Modifications
-------------
This changeset makes a couple modification which all have
the goal of improving the situation where a LDAP enabled
setup is used.

 - The SaslAuthHandler needs to remove itself from the pipeline
   when auth failed. This will make sure that later pipelined
   results are sent to their corresponding handlers and are
   not treated in the wrong place. It is not harmful since we
   are already in a bad state, but it makes debugging harder
   since you end up looking in the wrong place first.

 - When an authentication failed event is raised, we now include
   a more descriptive message that indicates potential causes,
   especially focussing on the (unobvious) LDAP problem.

 - The PasswordAuthenticator has been extended with a builder
   method and a static method that makes it much easier to
   "just" enable PLAIN which is needed when the server is !=
   6.5

Result
------
Better user experience when using LDAP and non-TLS connections.

Change-Id: I1cdc60a037ef1eee94384a207855c2edd4c0766f
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/136050
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>