JVMCBC-892: Check idle pool times way more frequently
Motivation
----------
Various couchbase server services have reduced their idle times
to 5 seconds to help mitigate potential slowloris attacks. This
unfortunately also has a profound impact on how quickly we need
to clean up idle endpoints, otherwise the server terminates
the sockets for us and we have noisy logs and lots of reconnect
attempts.

Modifications
-------------
The default idle timeout has been reduced to 4.5s in a previous
changeset, but when testing with recent 7.0 builds I noticed that
connections were still terminated on us. Problem is, the actual
timer that checks all endpoints on a periodical basis still had
a much higher granularity (5s!). In empirical testing, moving
it down to 500 millis looks to be granular enough to catch any
idle endpoint and disconnect it before the chance gets a server
to do so.

Note that I also modified the event we raise if the duration
is set too high to reflect the new reality.

Change-Id: Ie5c4b6b8bf73e75a08c6ce4ffba3225c8ae08a59
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/136519
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/137242
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>