JCBC-1717: Support for targeting requests to a given node
As part of the transactions-N1QL integration project, it's
necessary to always send QueryRequests related to a given
transaction to the same query node.

The majority of this work is being implemented in the
transactions library with a new TargetedQueryRequest.  All
that's required from the SDK is this change - adding
the last node used for a request into the RequestContext.

Change-Id: Ieeb7b87b4ba40753f013a95bdbd888832061075f
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/136873
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>