JVMCBC-892: Set idle time check to 100ms
A previous changeset set it to 500ms, but unfortunately in testing
I still saw occasional races with the server (given that we set it
to 4.5s and the server has it at 5s, if our interval is 500ms it
might not always work out).

At 100ms, I havent's seen any races so far.

Change-Id: Iee25c5242a0672bc7dfadcb2105a16be8db2a770
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/137138
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/137248
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>