JVMCBC-896: Fast Dispatch Pooled Requests
Motivation
----------
In the current pool implementation, if there is no available
endpoint but the pools upper limit is not reached yet we do
two things:

 1) open a new endpoint

 2) retry the request so that it gets a new chance of being
    dispatched to this or another endpoint

While this is simple and works in most cases, a couple of
issues have been identified with the approach:

 - If the above is true, each request gets at least retried
   once, which adds a minimum delay of 1ms to the request.
   That's not a lot, but depending on the situation it can
   be retried more often (see next case)

 - There are pathological cases where a request might not
   get handled for a long time or at all. If it happens
   to be the poor one which always opens a new connection
   and then another request steals the endpoint again it
   will be retried over and over again.

Also, one of the primary reasons the code worked like this
is that it was designed to be resilient against socket
connect failures. But it turns out, this penalizes the
good path (which will be way more common) for the case
where the endpoint is not connectable.

Modifications
-------------
Based on the observations above this changeset makes a
couplee important modifications (and actually brings it
way closer in-line with what we have in SDK 2).

The most important change is that if an endpoint can
be opened in the pool it is now "reserved" for the
request in question. The request will wait for its
endpoint and then two things can happen:

 - The endpoint is connected: send the request into
   it immediately.
 - The endpoint failed to connect: send the request
   into the retry handler.

While this means that in the error case the request
will take a little longer  to be retried, the common
good case will be faster.

Change-Id: I28f79e7d75653654a12b69d94dc28d3d9526dd8f
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/137144
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
(cherry picked from commit a842ed3934d8c05e259dbe91435c7d863bd4159b)
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/137212