JVMCBC-898: Add fallback for lastDispatchedTo
If the request has never been sent into the channel, thee
HostAndPort is likely not going to be set. But at a higher level
we also set the NodeIdentifier which can at least give us an
indication of the target host name. We loose the port, but for
debugging purposes we at least now where the op should've been
dispatched to last, very important for KV debugging.

Change-Id: I9c28af268ef21fb9d6305d6abb9d18e71eb86ec4
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/138627
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/139260
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>