JCBC-1731 (1/n): Refactor Tracing Interfaces
This changeset is the start of likely many commits which gradually
rework and extend the codebase to align with the new RTO v2 RFC.

In this change, the InternalSpan is gone, which naturally touches
a lot of places - but it is mostly just a refactoring. Where the
InternalSpan previously was used, there is now only a RequestSpan
used, which makes the implementation conceptually simpler.

Most importantly, no public API has been changed for now. Some
functionality on the opentracing and opentelemetry module has
been removed which will be re-added in a controlled way later.

Change-Id: I2fb0001a70418976cbb6c187c0832f71e1df22e7
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/139494
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>