JCBC-1731 (2/n): Rework JSON output
This changeset aligns the threshold reporter and the orphan
reporter with the new RFC, unifying the output format and
making it more consistent.

Also, this adds a "request name" to each request which mostly
can be derived from the name but not always, and where it cannot
it is just overridden  manually for simplicity reasons.

NOTE: since the target format is not yet 100% set in stone, it
is hidden behind a feature flag for now. Once stabilized, it will
be turned on by default.

Change-Id: I1507f8fe75ad604f9b2f8031a1658bda103d4dd5
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/139821
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>