SCBC-241: Refactor Document Expiry Duration
Allows the user to specify an expiry(Instant) as well as expiry(Duration)
on all relevant Options blocks.

As in Java, an expiry(Instant) is treated as an epoch time, while
an expiry(Duration) is treated as a relative time.  It's added to the
current epoch time before sending to the server, e.g. all expiry
times sent from the SDK are now absolute epoch times.

Not ported from Java, is the temporary special case logic code that
converts an expiry(Duration).  As this code is due to be removed in
the 3.1 release anyway.

Change-Id: I60788cd42f72f43acefd7a39b0368fbf02f34c79
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/140645
Tested-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>