JVMCBC-904: Dependency Bump and Test Polish
This changeset bumps dependencies before the release, but also
polishes some of the tests so that the warnings are not logged
but rather asserted instead.

Also note that one retry test has been removed because something
changed in the reactor code on the update. Since this is deprecated
code and just there for compat reasons I did not see any reason
not to remove this test.

Change-Id: I4df90864bbf6b32742c1369cce7f0d965f224190
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/141154
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>