JVMCBC-905: Allow to disable hostname verification
In SDK 2 we allow to enable hostname verification, in SDK 3 it is
enabled by default. In some cases, although it is less secure, it
must be possible to disable hostname verification in certain customer
environments.

This change allows to configure it optionally, but it will emit
a warning on startup to notify users in case they did not do
it intentionally.

Change-Id: I62d87753329233e2f6e9ae5b15661c074af72235
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/141239
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>