JCBC-1731 (4/n): Polish Metric Support
This changeset adds configuration, bumps up Otel to 0.11 and also
adds tests and cleans up all kinds of small things that I found
while actually using it in a sample project. This should get it
much closer to proper usage in the field, even though it is still
disabled by default to not interrupt anything.

Change-Id: Ic79968a7d58bd1745924f3e239dacdd6324709eb
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/141362
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>