JVMCBC-906: Remove Tracer argument from OpenTelemetrySpan
The tracer is actually not needed inside to complete the span, so
it can be removed as a mandatory argument, making it much simpler
to call from the user side.

Technically this is a breaking change but our API is marked as
volatile and still in progress, so it should be fine.

Change-Id: Ie10f63405e51016c225ad0ec90515c4a558ae0ab
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/141623
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>