JCBC-1747: Pass query_context to prepare and execute statements
A bug has been found in situational testing where adhoc queries
work fine but non-adhoc would fail. The issue comes down to the
fact that we are not passing down the query_context to the individual
prepare and/or execute statements, in which case the query engine
things we want to talk to the bucket.

Change-Id: I00bb7bc5e65c472dbdea79dc69840c41abc890ac
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/142059
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>