JVMCBC-909: Retry bucket open if not found until shutdown
Motivation
----------
Before this changeset, when a bucket was not found on the cluster
the open bucket attempt would be aborted and not continued. This goes
against the contract where we would continue to try and open it
until the cluster is disconnected or the bucket is created and
subsequently opened.

Modifications
-------------
This changeset glues together a couple pieces to make sure that
even if the bucket is not found the core will continue to retry
and - once the bucket is eventually created - give it a chance
to be opened properly.

1) If the cluster manager loader returns a 404, raise a
   BucketNotFoundDuringLoadException.

2) This will signal to the DefaultConfigurationProvider that
   it should retry the open bucket sequence (with a 1 second pause
   so we are not spamming the cluster).

3) The SelectBucketFailedEvent has been polished to make it
   clearer why it failed in case of an access error.

4) An integration test has been added to make sure the config
   provider keeps going in case of a bucket not found.

Result
------
Even if a bucket is created after the SDK is bootstrapped, it will
detect and pick it up properly.

Change-Id: I40de5af5ae75eef7d64ad29c42203b7055a78eae
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/142757
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>