JCBC-1730: Add collection support for BatchHelper
This changeset adds support for collections to the BatchHelper. It
does this by adding it ot the multi observe via cas request. Along
the way I extracted the collection encode method into a separate
one where the key can be provided explicitly.

Also, the LEB skip method now returns the number of skipped bytes
so it can be used when calculating the key size.

Change-Id: I7eff2ec6a4ebe1a5d05706fc6900364105e3e832
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/142974
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>