JVMCBC-916: Use remote hostname for HTTP host header
Motivation
----------
Right now we extract the hostname for the HTTP host header
on HTTP request out of the InetSocketAddress from netty, but
we always use the IP address. This works fine in most cases,
but in others it can be confusing (i.e. where hostnames are
used throughout the system including the server config).

Modifications
-------------
Since we already have the remote hostname string in scope from
the endpoint, we can just make use of it directly. We do the
same for the port.

Result
------
The HTTP host header now reflects the string used in the server
config and is consistent throughout the system.

Change-Id: Iac99b8c0fa80ffb615dbfba50db0001058147a70
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/144361
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>