JVMCBC-920 Move MutationState logic to core
Motivation
---------
Share it with Kotlin (and maybe Scala?)

Modifications
-------------
Add a new `MutationTokenAggregator` to core-io.
This class works with Maps instead of the java client's
JsonObject, so it can be shared between JVM SDKs.

Store the tokens in a ConcurrentHashMap so tokens
can safely be added by different threads.

Index the tokens using bucket name + partition
as the map key so adding a token is more efficient.

Modify Java client's MutationState so it delegates
to MutationTokenAggregator.

Change-Id: I4fd2b995309a5fb3b0d8050f4fae9e05e33d95f9
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/145339
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>