gardening: regularly print sample bucket json docs
The GetResult toString() code tries to make a decision if it is a
printable string or binary data. It does this by checking the flags,
but the server sample buckets do not set the common flags. As an
easy workaround, check if the flag is 0 AND the first char starts
with a {.

Change-Id: I7635180cc60ffba13d17b445ea3d776c83bed576
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/146684
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>