JVMCBC-921: 5/n make sure get_all_replicas is not completed too early.
This changeset makes sure that the get_all_replicas parent span is
only completed when all the inner completable futures are done, not
just when they are created (which would be way too early).

As a drive-by change, the tracing identifier for get_any_replica
is stored in the TracingIdentifiers class.

Change-Id: I6a33bdde8df5b47d003fc42ec399240716cb1bf4
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/146685
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>