JCBC-1775: Mark GetAllDataverses as uncomitted.
This API is only available in java, and until all SDKs have it
and it is part of the RFC, mark it as uncomitted. It is unlikely
to change at this point, but let's be accurate.

Also, the API has been missing options and a proper tracing span,
so adding it as well.

Change-Id: I025f87d9d673130f64fa1286de7b6e5e46ab1ad1
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/146710
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>