JVMCBC-936: Allow to access list of default ciphers.
The previous changeset added the capabilities to provide a custom
list of ciphers which is helpful on its own if the list provided
should be very explicit.

If the use case is to _add_ more ciphers to the default list, it
can be a little tricky to extract the default ciphers manually
and then hardcode it.

This changeset adds static helper to extract the default list, which
makes it very convenient to add more ciphers to that list before
passing it into the ciphers method again.

Change-Id: Iffac76e35968118c4172679072634c9d01dd48f4
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/147901
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: David Nault <david.nault@couchbase.com>