JVMCBC-951 Move GetAny/AllReplicas into core
Motivation
----------
Share it with other JVM clients.

Modifications
-------------
Add ReplicaHelper to core, with code moved from
AsyncConnection and ReactiveCollection.

Add GetReplicaResponse to core. This is what ReplicaHelper
returns. Clients convert it to their native representation.

Update the ReactiveCollection javadoc to reflect the
actual behavior (which is not changed by this commit).

Add integration tests for the reactive implementation.
The tests pass before and after this commit.

Change-Id: I87e37d04f2f786e485ee053392d3f17966463dbf
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/149122
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>