JVMCBC-954: Use FF map only if NMVB at least present once
Motivation
----------
Before this change, during a rebalance (if a FF map is present
in the server config), the client would try the fast forward
node even if the operation got retried for a different reason
other than a NMVB in the first place.

While it is true that moste of the time in this case it could
be a NMVB, it doesn't have to be all the time. Since the server
expects us to ONLY try an operation on the FF node if a NMVB
got returned, the logic needs to be adjusted.

Modifications
-------------
Instead of checking on the number of retries (> 0), this code
change now increases a couinter on the request if a NMVB response has
been received and only then it will try to attempt to use the
fast forward map.

This information is also shown in the service context so in the
future we should also see if a request got rejected as part of
the error.

Change-Id: I66c005ec7b23c8c26a6268636eb8a1595d1da27f
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/150085
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>