JVMCBC-958: Do not delay individual collection id fetch
Motivation
----------
During code review it has been found that when a collection id
is not found for a scope/collection combination AND a fetch is
currently in progress, the operation will be retried.

This logic can be improved because right now it will even retry
if the id fetch for a different collection is in progress, potentially
serializing the lookups which can reduce performance if one
opens many collections at a time.

Modifications
-------------
This changeset contains:

 - The internal set is now concurrent and allows for a more
   fine grained control, which is reflected on the exposed
   APIs on the config provider.

 - Add a event that allows better debugging of outdated
   collection ids.

 - Also, one RetryReason has been moved from always retry
   to not always retry to align all of the collection outdated
   related retry reasons.

Change-Id: I916fab05e2419c9417fdbd8121b2a3c80d1cb5f9
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/150959
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>