JVMCBC-962: Add (internal) request callback functionality.
This changset adds request callback functionality, for now internal.
One callback is added that is executed before send in core. It is
used by the MCA client to attach to sent request and detect failed
operations in the callback.

This infrastructure will also allow to add more callbacks in the future,
i.e. for auditing purposes where needed.

Change-Id: I2aa22d7c4f381783cbdc4c9b8bbdeee11eadcd22
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/151523
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>