JVMCBC-963: Retry more bucket and global config load failures
Motivation
----------
It has been found that in certain situations the client will
back out of the loading process even though it can and
should keep retrying (for example if all nodes are down while
the client bootstraps and they come online later).

Modifications
-------------
This changeset modifies the error/retry handling behavior of
the config provider in the sense that it, instead of being
opt-in for what should be retried, it is now opt-out in what
should not be retried. This will make sure that all kinds of
(transient) errors can be remedied once the situation in
a different component resolves itself.

Result
------
The bootstrap process is now more resilient to temporary
conditions like all nodes being offline when the client
is initiated.

Change-Id: I6a091e27af68a716f0fb657f98c031fc2247fe13
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/151615
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>