JCBC-1804 Ensure contentAsObject/Array() works with custom JsonSerializer
Motivation
----------
contentAsObject() and contentAsArray() depend on the JsonSerializer.
If the user provided a custom serializer that did not know how to handle
JsonObject and JsonArray, these methods would fail.

Modifications
-------------
If the user provides a custom serializer when building the
cluster environment, wrap it with a serializer that knows
how to handle JsonObject and JsonArray.

Change-Id: I39b0fe02e8d9043e41af7cb2cde42c6a63e6abf4
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/151771
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>