JVMCBC-967: Fix OverflowException with PersistTo/ReplicateTo
Motivation
----------
Inside the observe functionality, the SDK needs to "keep polling" when
the desired state is not yet reached. It does this with an exponential
backoff and a repeatWhen reactive operator. Turns out that with the
initial small delays we are using there could be a race condition in
the reactor Mono.delay operator which leads to an OverflowException.

Modifications
-------------
There is not much we can do about the overflow (the reactor team will
investigate that separately), but together with the all ever helpful
@simonbasle we were able to craft this workaround that ignores the
issue at hand and as a result makes sure it eventually succeeds. Since
it's a race, after ignoring and resubscribing again eventually it
will work.

Result
------
No more OverflowException when PersistTo/ReplicateTo (observe) is
used.

Change-Id: I298f7c19a3551af147fb4d8242794805dd7150ed
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/152087
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>