JVMCBC-997: Duplicating attributes from dispatch_to_server to improve debugging
The full justification for this change is in an RFC comment.  The short
version is:

Most, perhaps all, current tracing systems don't allow performing
queries across spans.  E.g. there is no way of finding
dispatch_to_server spans where <some attribute in a parent span> ==
<some value>.  One useful query along those lines would be
checking the server_duration times against different buckets.

After discussions with one of the tracing providers, we understand
there are two solutions to this: duplicate some attributes between
spans, or use OpenTelemetry 'baggage'.  In practice, baggage does not
seem useful: it is unclear/undocumented how to attach it through the
API and, many (perhaps all) tracing systems do not appear to expose it.

So, this changeset duplicates some useful attributes between the
dispatch_to_server and top-level KV span.  The justification for these
attributes in particular is included in a RFC comment linked to from
the parent ticket.

It would also be helpful to have the operation type ("upsert") available
on the dispatch_to_server span, but this would be a larger change.

Change-Id: I1499de0030f48e5186c58409ba963746d55fa357
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/152452
Tested-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>