JVMCBC-969: Add core infrastructure for the eventing service.
This changeset adds the MVP code for the eventing service, it
does not even add a request/response that runs through. This will
be done in a follow-up commit when the java API is added on
top.

Note that on purpose the eventing service is not tracked in the
threshold and orphan tracer, similar to the management API.

Change-Id: I1fe501db70d1433f5ac089754bc1e4c5ec939b85
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/152690
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>