JVMCBC-977: Adds support for revEpoch to core-io.
This changeset adds support and tests for the new revEpoch
feature which the server sends us in the config from 7.0 forwards.

Note that the epoch (if present) starts with 1, so we assume 0
as the default for all older configs which do not have an epoch.
Having an epoch is always better than no epoch, and if the epochs
are the same (including 0) we still fall back to just checking
the rev.

Think: the epoch and the rev act like a major and minor release
when comparing. So a higher epoch and lower rev always trumps
a lower epoch and higher rev.

Change-Id: I25a90d7b6cd035a5572d65c01311e52dd90b70a5
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/154566
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>