JVMCBC-979: Rename ThresholdRequestTracer to ThresholdLoggingTracer
This changeset renames the tracer based on recent RFC changes, but since
the config is marked as comitted it is still possible to pass it in the
environment as before. Internal logic has been added so that if the
deprecated config is passed in, it will be converted to the new one
and vice versa.

Change-Id: I28868f23a8077f938589794fb54aee13aab4fad1
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/154745
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>