JVMCBC-987: Allow to provide SDK2-Compat Memcached Hashing Strategy
This changeset adds support for a SDK2-compatible hashing strategy
for memcached buckets. This is important if users need to upgrade
from SDK 2 and do not want to partially invalidate their caches.

Note that it is only recommended for users upgrading, since the
default sdk 2 strategy is NOT compatible with libcouchbase ones.

Also, it has the downside that it potentially performs reverse
DNS lookups which is something we avoid completely in SDK 3
otherwise.

Change-Id: Ib14bf5fae8248dbfe4046a07baecfa643ebfa275
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/155687
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/158047
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>