JCBC-1787 (Followup) Validate expiry instants
Motivation
----------
An earlier change during development of 3.2.0
made it invalid to pass Duration.ZERO as a document expiry,
but this caused problems for the Kafka connector.
It's reasonable to assume there's other code in the wild
that depends on Duration.ZERO meaning "never expire".

Modifications
-------------
Once again allow Duration.ZERO to mean "never expire"
(instead of being an invalid argument).

The prohibition of Duration.ZERO never made it into
a released version of the SDK, so this change shouldn't
affect developers.

Change-Id: I08fe79d59c6ac52381954185f673cfa37e8e9b7f
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/156205
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>